Move OTP functions into helpers
